avocado.multiplexer: Modify multiplexation mechanism
Instead of detecting non-sibling nodes generate list of leaves per
each parent and pass it up the tree to the root. Restuls should be
the same as before, only speed is improved.

Additionally allow specifying nodes (!join), whose children belongs to
the same group. This extends the possibilities beyond what was possible
before.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>