avocado: Add support for recursively finding tests in a directory
Add support to inspect and find tests inside a directory, recursively,
by using the method `discover_url`. It takes a url (path) and
returns the tests parameters it discovers.

The loader object has the .discover() method, that can be used
to inspect the location, and the .validate() method, that will
discover any problems on the user's input.

We are also changing the behavior of avocado to error out on
invalid inputs (missing paths or files that are not avocado
tests). Therefore, the concept of MISSING or NOT_A_TEST both
disappear from the UI. Unittests were updated to reflect the
new status quo.

Signed-off-by: Rudá Moura <rmoura@redhat.com>
Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>