rest-client: add initial module namespace and response handling module
An essential part of a client interacting with avocado-server is to handle
the reponse of a request. The most common type of response type is a
results reponse, when one performs a GET to a resource "root" path and
and expects to return a list of resources.

To make that more pythonic, this commit introduces a ResultResponse class
that accepts the JSON response data and checks for completeness. If
results response fails to comply with the required data members, meaningful
exceptions will be raised.

Signed-off-by: Cleber Rosa <crosa@redhat.com>