avocado.plugins.remote: Disable timeout for remote execution
Default timeout for execution is 60s. This patch removes this for the
remote `avocado ...` run as there is no sensible default for it, nor
convinient way to set it on the cmdline. It should be relatively safe as
Avocado on remote machine should timeout using the same timeouts as on
host and in case the machine dies fabric fails the execution and
returns.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>