Command line parsing: do the command line test list parsing in argparse
Instead of delaying the parsing of the test list to code that handles
the jobs, let's handle that early, and in the usual Python way, simply
by relying on argparse.

For other, possibly more complex command line requirements, custom code
can be written, but it should still be placed at the argparse level.

Signed-off-by: Cleber Rosa <crosa@redhat.com>