docs: Remove docs/build from the repository
As far as I can see and test, the directory doesn't have to be present in the
repository, as it'll be created on-the-fly during the documentation generation.

As a bonus, avoid a dangerous "rm -rf $(VARIABLE)/*" in make clean (which is a
pattern for catastrophic results -- if $VARIABLE, whatever it is, is not set,
this will result in "rm -rf /*").