avocado: Paginator changes
Some plugins don't really need to be started in paginator
mode (such as the test runner). Only plugins that need to
list things as opposed to execute test operations will
need it. Therefore:

1) Remove use_paginator=True from test runner, remote
   and vm plugins
2) Add use_paginator=True for the test lister plugin
3) Fix bug in paginator use that will make echo to
   be turned off if you use the paginator (causing
   input to not be displayed in your terminal after
   you use 'avocado list', for example).

The problem with the paginator bug is that it is
explicitly necessary to close the paginator once
we are done with it, as opposed to let it get
garbage collected. For that, we introduced the method
View.cleanup().

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>