avocado.sysinfo: Introduce system information profilers.
Sysinfo module now allows to define profilers, which are programs
defined in command line, to start as a background processes and log
the output in a proper directory. Example: vmstat 1.

The profilers are started before a job execution and then, stopped at
the end of the job. The output is stored inside `sysinfo/profile/*`
subdirectory, in job result directory.

The profilers can be customized in `avocado.conf`, section
`sysinfo.collect`.  The default configuration define
profilers as the following: `journalctl -f` and `vmstat 1`.

This feature is turned off by default.

Signed-off-by: Rudá Moura <rmoura@redhat.com>