avocado.test: Upon serialization, sanitize the test exception class
Sometimes, exception classes might be written in a funky way,
and upon unserialization their __init__ methods might be
called. In this case, if the __init__ gets called, the
needed arguments won't be passed and then avocado will crash.

As we don't really need the actual class, just pick the
string representation of it.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>