avodado/job.py: implement job standalone mode
The idea of the standalone mode is to be able to have a avocado test to
behave like a regular script, outputing to STDOUT the basic log (the
same as --show-job-log would) and return an meaningful status code). This
should help avocado tests to be run along any other test code or framework.

In this mode, a job will still be created, and the basic output files will
also be created, but will be hidden from the user, and removed after the
job finishes running.

This is a partial solution to the so called "jobless" mode, which would
allow a test to run without a job.

Signed-off-by: Cleber Rosa <crosa@redhat.com>