avocado/utils/path.py: add default parameter to find_command()
It's a very frequent code pattern to look for a binary in the system
using find_command(), and, if no command is found, trap the exception
generated and return a fallback.

This adds a `default` parameter that makes it possible to follow that
pattern without code outside the function itself.

Signed-off-by: Cleber Rosa <crosa@redhat.com>