fiotest: update the engine name for async I/O
It looks like the name of the loadable engine on fio itself has changed
from simpler "aio" to either a Linux or POSIX version.

Based on my tests on build attempts, it looks like POSIX is more dependant
when it comes to building out of the box. Let's default on that version.

This fixes the following error when running fio:

 fio: engine aio not loadable
 fio: failed to load engine aio
 ...
 fio: file:ioengines.c:99, func=dlopen, error=aio: cannot open shared object file: No such file or directory

That resulted in a test failure.

Signed-off-by: Cleber Rosa <crosa@redhat.com>