avocado-run-testplan: introduce a minimalist tool to help with manual test plans
Some test routines can not yet be easily automated in Avocado, and for those, we
rather test things manually then not at all.

The idea behind this tool is to use a predefined test plan (a simple JSON file)
that has test names and descriptions with how to perform the test and what to
expect. The tool will ask for the results for each test (either PASS, FAIL or
SKIP) and possibly additional notes. These results will then be saved to a JSON
file or to a human readable report.

The human readable report can, for instance, be part of a release commit message,
or if preferred, the JSON result file can be committed to keep track of what is
working and what is known to be broken at a given time.

Signed-off-by: Cleber Rosa <crosa@redhat.com>