avocado.core.data_dir: Rename get_job_logs_dir to create_job_logs_dir.
Since the API creates a directory (instead of just return the name),
rename `get_job_logs_dir` to `create_job_log_dir`.

Use a proper parameter to get the base dir, instead of passing
the whole `args` environment for that.

As a matter of style, `avocado.core` should avoid to use `args`.

Signed-off-by: Rudá Moura <rmoura@redhat.com>