avocado.job: Job class will only use args.
From the Job class, put the previous `standalone` argument
inside the `args` environment, so the class now requires only
one argument args.

For completeness, the Job class will always contain an `args` defined,
even if we don't pass anything, it will contain an empty environment,
so that we don't need to check if job is None and it will always be
possible to use `getattr()` to get an attribute from `args`.

Signed-off-by: Rudá Moura <rmoura@redhat.com>