avocado.test: Escape filename when executing SimpleTest
Current behavior of process.run(shell=False) is it expects multiple
arguments and split the provided cmd. This doesn't fit the SimpleTest
as it founds a binary and then tries to execute it without any
arguments as it might be separated by spaces. This patch escapes the
executed command using "pipes.quote" to avoid it's splitting.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>