avocado/job.py: disable local test file checks when running on remote machines
When using `--remote-no-copy` the test files will neither be copied from the
current system nor executed on it, so disable all sorts of checks.

This is a fix for the local check only. A more complicated issue is to perform
similar checks on the RemoteTestRunner and report possible failures in a clean
way.

This address the github issue #417.

Signed-off-by: Cleber Rosa <crosa@redhat.com>