Remote test support: add check for tests before trying to run them
Adding this extra check will give more precise error messages to the user in
case the given tests do not exist on the remote system. This is basically a
safeguard against a user using `--remote-no-copy` and asking to run tests
that do not exist on the remote system.

When not using `--remote-no-copy` the tests will be copied and this check
should always succeed. But, because of timing issues, concurrent access to
the remote system, and so many other issues could result in tests being
removed, I believe the user can still benefit from this (lighweight) check.

Signed-off-by: Cleber Rosa <crosa@redhat.com>