pyliblzma: make it optional since it is not essential
The line above says enough, but let me stress the point that we want
Avocado to be portable enough to run on various and possibly limited
systems.

Embedded systems and lightweight VMs/containers are a very compeling
target for avocado, so IMHO we should strive to make the avocado core
runnable with as little as a standard Python installation.

Signed-off-by: Cleber Rosa <crosa@redhat.com>