.travis.yml: Add python 2.6 among the testing targets
Add a new python 2.6 target so that we can make sure
avocado works on older distros (our specific focus here
is RHEL 6, but of course, we expect the work on RHEL 6
to be useful to other distros sticking with older versions
of the python runtime).

In order to do that, we added a new 2.6 related requirements
file, and changed the requirements a bit so that it all
works under the particular Travis CI environment.

Changes from v1:
 * Per ldoktor's suggestion, use conditional install of backports
   using $TRAVIS_PYTHON_VERSION

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>