.travis.yml: Add python 2.6 among the testing targets
Attention: This is *not* expected to pass on the 1st try.

Add a new python 2.6 target so that we can make sure
avocado works on older distros (our specific focus here
is RHEL 6, but of course, we expect the work on RHEL 6
to be useful to other distros sticking with older versions
of the python runtime).

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>