avocado/utils/io.py: rename to genio.py (meaning generic I/O)
The reason for this rename is that there's a name clash with Python's
standard library `io` module. This name clash manifest while performing
a simple Python validation of the avocado/utils/archive.py module:

$ python avocado/utils/archive.py
Traceback (most recent call last):
  File "archive.py", line 22, in <module>
    import zipfile
  File "/usr/lib64/python2.7/zipfile.py", line 501, in <module>
    class ZipExtFile(io.BufferedIOBase):
AttributeError: 'module' object has no attribute 'BufferedIOBase'

The explanation to that error is that Python itself adds the base
directory where the "executed" module file lives to the head of
the module search path list.

After some dicussion we realized that we should strive to not
name modules with the same names as Python's standard ones.

Signed-off-by: Cleber Rosa <crosa@redhat.com>