avocado.plugins: Fix #486 - Add option --disable-paginator to list commands
Turns out that depending on guest config options, sometimes
the paginator kicks in and won't release the stdout until
a human comes in and presses 'q'. This is very, very bad for
automated procedures such as the one used by the remote plugin.

So add --disable-paginator in both plugins that use that
functionality, so they can be used in automated scripts.
Afterwards, we'll use that option in the remote plugin.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>