avocado.test: Always get unique name in get_tagged_name
When no tag is supplied get_tagged_name() generates unique name. On the
other hand when tag is supplied, it silently reuses existing path. This
patch adds check for these cases and use $name.$tag.$number in case
$name.$tag exists.

This should avoid clashes when executing multiple tests with the same
name with multiplexation.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>