WIP: avocado.plugins: Add virt_test compatibility plugin
This is a very rough through the edges hack that makes
an avocado plugin that implements the virt-test runner.
It accepts the same params that the original ./run
application does, so you can do something like:

avocado vt-run -t qemu --bootstrap
avocado vt-run -t qemu

And it will work. Next steps are to make things to run
in the avocado main loop, instead of virt-test's.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>