avocado.plugins.htmlresult: Entirely remove --relative-links
The option --relative links instructed the htmlresult plugin
whether to use absolute log paths in anchor tags, or to use
relative paths, something that dates back to the old autotest
html job report. After thinking about it I realized that this
option doesn't need to be here, as using relative links is
basically better for all intents and purposes (the job results
can be safely redistributed and the report would still work).

So get rid of that option and always use relative links for
the job results dir links and logs. Tested with local execution
as well as remote/vm execution.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>