avocado.aexpect: Fix bugs that broke avocado-virt
We can't require avocado imports in the server part of
the aexpect process because that will be executed
stand alone, and will not have access to the rest
of the avocado libraries. The bugs were introduced
in commits:

6f380e0d77d839dac1b30987e84fa300b4a99058
3fa9077f133b2445f839056cd568d21d14f7975d

This patch fixes the problem by moving the imports
to the client side of the module.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>