avocado.spec: use the Makefile target for man pages
This means that we include the Makefile in the source tarball and just
use the `man` target.

With regards to portability among Python version, this should not be
an issue as rst2man on the system itself will the correct version. In
a sense, it's just another utility, it doesn't really matter it's
written in Python.

Signed-off-by: Cleber Rosa <crosa@redhat.com>