unittests: run them during the build of RPM packages
It seems a good idea, and another safety net, to run the selftests
during the build of packages. For now, it's not trivial to run the
functional tests, so let's at least run the unittests.

Another current limitation is that unittests are only run on Fedora
and not on EL6 or EL7, and the reason for that are missing
dependencies on both EPEL6 and EPEL7.

Signed-off-by: Cleber Rosa <crosa@redhat.com>