avocado.spec: reorganized runtime and build requirements
The basic idea here is to put all common requirements on a common line
and deal with the differences in conditional blocks.

Also, python-flexmock is not available on EPEL 7, but available on all
other supported platforms.

Finally, python-yaml and PyYAML are in their own lines because it
makes it easier to see that they're the same package with different
names.

Signed-off-by: Cleber Rosa <crosa@redhat.com>