avocado.multiplexer: Improve the params clash detection
Previously we only detected the returned value. This patch adds real
params clash detection as it tracks the origin of the param and in case
it's not the same it raises exception.

The difference from previous version is that even when we define the
same value, this version detects it's from different source and raises
exception.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>