sysinfo: replace hard coded commands and files for configurable files
This change adds flexibility to the current sysinfo implementation in
the sense that the commands to be executed and files to be collected
are no longer hard coded, but configurable.

The configuration file gains a section for that, and the files
themselves are plain text files that can be easily customized by users
to include extra commands/files.

Also, the the configuration for the profilers has been adjusted for a
uniform look, feel and experience.

Signed-off-by: Cleber Rosa <crosa@redhat.com>