sysinfo: replace the term loggable with collectible
First, this is purely a name change, no change of behaviour should
happen because of this.

Second, this is a naming issue, which means it can sound like a great
or awful idea and both opinions would be right about it.

Finally, IMHO, the name collectible is better suited here for a couple
of reasons:

 * It breaks apart from the already poluted use of log, logfile,
   logging, etc
 * It does not suggest that something will be logged in the
   traditional sense, that is, it will not go through the `logging`
   mechanism
 * It's more closely related to what us humans want, collect the files
   or command outputs for later analysis

Signed-off-by: Cleber Rosa <crosa@redhat.com>