avocado.loader: Fortify test discovery against stdin/out/err corruption
There are python traps (they can't be called scripts) which on import
modify stdin/out/err. This even on discovery makes avocado behave very
badly and it's hard to discover what and when it happened. This patch
provides dummy stdin/out/err when discovering the tests.

This solution only takes care of stdin/out/err related issues, safer
method would be to spawn another process and gather the info there.
Still we can't avoid modifying the OS during the discovery, but that's
IMO acceptable for a test-framework.

The examples/tests/failtest_ugly.py is used in basic_tests unittest
using "avocado list" (without this patch it waits for input, than it
prints the test's outputs and exits without listing the tests).

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>