docs: add section on GDB support and avocado-virt usage
It was attempted to fill in the gaps so that avocado-virt would be
able to transparently use avocado GDB support. It turns out that a
number of foundation changes will have to be made to support launching
process in the background while running inside GDB.

So, for now, let's document that limitation.

Signed-off-by: Cleber Rosa <crosa@redhat.com>