Makefile: replace source tarball generation from distutils to git
It was noted that, on some Python version, the generated tarballs
actually have a lot of files missing. IMHO, if we track a file on git,
then it should make into the tarball. Also IMHO, the distutils sdist
code is a nice thing to have if you can't export your code right from
your SCM.

This is kept simple on purpose so that if we hit any unintended side
effect we can revert to the previous way of generation source
tarballs.

Signed-off-by: Cleber Rosa <crosa@redhat.com>