Unittest fixes: Problems with loading tests from multiple sources
When loading tests from multiple sources, we have a dillemma: what
to do with tests that don't resolve to a filesystem location, since
future loader plugins might be able to find it.

This is a workaround, and helps some unittests to pass. But
as a matter of fact, there are some important unanswered questions
with regards to the behavior of avocado when different inputs
are passed to it.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>