avocado.plugins: Store failed plugins instead of logging them to stderr
Store failed plugins instead of logging them to stderr.
That will allow us to give better information to the
user as of why a given plugin was disabled.

This also introduces the 'disable_reason' attribute to plugins,
that lets us to explain to the user why one of them might be
disabled.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>