avocado.multiplexer: Make test sources able to merge mux params
Some upcoming test sources may want to simply append
multiplexer params to their original params, instead
of having them overwritten by the muxer. Define a
mechanism to do that. Test sources, as they assemble
the params that will be used with the test classes, may
define the key avocado_mux_merge and set it to True.

If the muxer finds out about that tree, it will append
the params from the muxer in the param 'avocado_mux',
making it possible for said upcoming test classes to
retrieve avocado multiplexer params.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>