avocado..tree: Overwrite multiplex only when explicitely asked for
Originally we used multiplex = True and False. This patch adds None and
the value is not updated during merge when the new node's multiplex is
None.

Currently this means that once node is set as multiplex, there is no way
to undo this. We might consider creating !nomux flag in the future to
force multiplex = False, but this is currently not needed.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>