trinity example test: fix fetch of tarball parameter and update to 1.5
Because of the parameter behaviour change this test was failing to get
the name of the tarball to use during the test.

Once that was fixed, there was a couple of build errors. Instead of
debugging the build of the 1.4 version, I tried version 1.5, which
built without any problems. Let's use this oppotunity to fix the
test (wrapper) and update the (original test) tarball.

Signed-off-by: Cleber Rosa <crosa@redhat.com>