API review: move avocado.aexpect to avocado.core.aexpect
Even though aexpect is possibly a good candidate for the avocado.utils
namespace, there are plans to actually move it into a standalone Python
project/module. So, since no tests are currently using the module
directly, let's move to avocado.core.

It can not be removed at this point because the avocado-virt extension
uses it. A PR is being sent to avocado-virt to sync with this change.

Signed-off-by: Cleber Rosa <crosa@redhat.com>