[RFC] API review: use relative imports in avocado.utils modules
It was mentioned that we should think of avocado.utils as a subproject
that happens to live in the same source tree. So, it should not make
use of any avocado module. If that's the case, maybe we don't want to
keep referencing 'avocado.utils' in the utils module themselves.

This is in accordance with PEP-0328[1]

[1] - https://www.python.org/dev/peps/pep-0328/

Signed-off-by: Cleber Rosa <crosa@redhat.com>