API review: move logging config specification and action to avocado.core.log
It looks like there's no need for the logging configuration and the
application of that configuration to live under avocado main module,
so let's move it to avocado.core.log.

Also, there also seems to be no need for the configuration to be
applied at the main module import time.

Since when the avocado command line test runner is run and also
when a test is run "directly" results in instantiating an
avocado.core.app.AvocadoApp object, this looks like the best place
for the logging configuration to actually be performed.

Signed-off-by: Cleber Rosa <crosa@redhat.com>