avocado.utils.software_manager: Check if yum module is present.
When using the software_manager component on a Red Hat like distro,
if the yum module for Python is not installed, log the information
as an error message and use the basic support from rpm/yum.

Why this?

If the module yum is not installed (for example, in a virtualenv),
the software_manager component will not traceback when creating
an instance of YumBackend and only the method provides() will be disabled.

Signed-off-by: Rudá Moura <rmoura@redhat.com>