avocado: Avoid loading plugin classes in the code based on hardcoded arg names
For a while we have been using code similar to this:

def load_plugins(self, args):
    for key in args.__dict__.keys():
        if key.endswith('_loader'):

One could argue that the sufix check is not quite elegant
or correct, so we can remove that and get straight to the
business of checking the actual loading class we're interested
in. This PR changes all the occurrences of that practice inside
the avocado code.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>