docs: simplify basic example on Writing Tests Guide
The idea is to remove everything that is not critical to the user's
understanding of the core concepts.

Also, to sync to the new multi-test in a single class capabilities
of Avocado, let's name the main test method `test`. This paves the
way for other (multi test) examples that name test methods with
the exact same prefix (test_foo, test_bar and so on).

Signed-off-by: Cleber Rosa <crosa@redhat.com>