avocado: Add a mechanism to skip notifying broken plugins on stderr
Implement a feature to skip notifying broken plugins on
stderr. Let's say a user has a broken plugin that can't
be fixed right now, and he/she doesn't want to be bothered
by having it notified in stderr all the time avocado is
executed, add a configuration option that allows you to
selectively ignore plugins to display errors in stderr.

The reason why the feature is configurable is that we
want users to be notified when a new plugin is present
and for some reason is also broken, this finer grained
control is more useful to users IMHO.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>