doc review: be clear about using any language or Python with libraries
First, the wording was not clear about the language choices and the
added benefits of using Python. Also, let's focus on using the term
libraries instead of APIs. Libraries are an important part of Avocado
(see definition of what Avocado is) and IMHO denote more value than
the term APIs does.

Signed-off-by: Cleber Rosa <crosa@redhat.com>