doc review: move result specification to the Reference Section
As this pretty much documents how Avocado saves the results, it's
pretty much a reference document.

The location of the job logs is pretty clear in the Getting Started
section when the user runs their first tests, and it's pretty obvious
even for 1st time users not even reading the documentation.

If users want to understand more about it, or customize the location,
the Reference Guide is indeed the place to go.

Signed-off-by: Cleber Rosa <crosa@redhat.com>