API review: allow tests to fail/skip/error themselves by using methods
Instead of playing with core exceptions or downgrading those exceptions
to the base Test API, let's allow users to do so with convenience
methods.

The example tests now make use of those, and can now remain in the
examples directory instead of the former proposal of turning them
into "functional test assets" and moving them to the selftests
directory.

Signed-off-by: Cleber Rosa <crosa@redhat.com>