avocado.core.remote.runner: When setup/teardown fails, exit gently.
When setup() or tear_down() fails, Avocado
will provide a gentle error message and a proper exit code and
put the traceback in the test log.

  ./scripts/avocado run --remote-host nowhereland passtest
  LOGIN      : rmoura@nowhereland:22
  Avocado job failed: JobError: Name lookup failed for nowhereland

Prior to this commit, when setup or tear_down failed,
Avocado dumps the traceback and prints that it crashed, which
can potentially confuses the user.

Signed-off-by: Rudá Moura <rmoura@redhat.com>