API review: implement relative imports also on avocado.core.*
The same approach has already been applied to avocado.utils itself.

The references to avocado.utils in avocado.core, though, were not
made relative. The reason is that, according to the PEP 328, going
up three levels while "legal, it's certainly discouraged"[1].

If we reach a consensus that we should do it all the way, then
another commit can deal with that. If not, we may even think about
making avocado.utils a library (repo?) by itself.

[1] - https://www.python.org/dev/peps/pep-0328/#guido-s-decision

Signed-off-by: Cleber Rosa <crosa@redhat.com>