avocado.core.test: Fix problem with unittest compatibility
Introduced with commit d6abdcd4a6ea382c0447ece49f983b37823de245,
avocado tests no longer report correctly to the unittest runner.
The reason is that there is a structure that must be correctly
updated by the run() method (the result instance), that was not
being updated in the new implementation.

Therefore, one possible fix, presented here, is to pick the
result instance and update it accordingly. The problem with this
is that we still don't reach full compatibility with unittest,
given that:

 * The behavior of the method is different from the unittest
   counterpart (the avocado implementation may raise exceptions,
   as opposed to the unittest implementation
 * The unittest implementation updates some other internal
   structures and does extra checks (for example, the skipped
   test check), meaning that we'll have to do further work on
   that method, and having an implementation that has to count
   on internal structures of a class is bound to have a problem,
   should the internal structure change. Since we are talking
   python 2.6/2.7, we are probably off the hook, given that we
   don't expect that code to go to any major changes.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>