avocado.core.test: Fix problem with unittest compatibility
Introduced with commit d6abdcd4a6ea382c0447ece49f983b37823de245,
avocado tests no longer report correctly to the unittest runner.
The reason is that there is a structure that must be correctly
updated by the run() method (the result instance), that was not
being updated in the new implementation.

Therefore, one possible fix, presented here, is to once again
*not* implement run() at all. this way we avoid messing around
with internal data structures, and when avocado is executed
under nosetests, or the default unittest runner, we get the
expected behavior of the method. The downside is that we have
less perceived integration among both classes.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>