API review: move runtime configuration to the avocado.utils
The runtime module had already been diminished by moving some of its
well, runtime configurations, directly to appropriate modules.
The still standing runtime configurations are: current test and
current job.

It's clearly *not* the best design to have common utility modules,
that is, avocado.utils with *knowledge* about the inner workings
of an external entity such as the Avocado Job and Test classes.
It would be much better for Avocado Job and Test classes to
register hooks/callbacks with the utility methods in a agnostic
way.

Since some of the utility modules were not designed as external
entities, the coupling is still pretty tight, and some more
profound API changes would be necessary.

For now, our goal is to avoid the crossing of boundaries when it
comes to importing and using modules, having a sane, layered approach.

The next level, still pending, is to remove this knowledge dependency
and make the utility code trully generic.

Signed-off-by: Cleber Rosa <crosa@redhat.com>