API review: move GDB configuration to the avocado.utils.gdb module itself
Instead of having a "common ground" that two namespaces would import and
share runtime information, let's just make the plugin set the defaults
for the module.

This is quite appropriate, both regarding the namespace where the
configurations are kept (in the avocado.utils.gdb module itself) and
also regarding the single direction of module layering.

Signed-off-by: Cleber Rosa <crosa@redhat.com>