API review: let the process API consumer set the exception for undefined behavior
One of the possibilities when Avocado delegates the execution of a process to
the utility library process is undefined behavior.

This can help, for example, when this library is set to run processes in more
complex environments, including with users manipulating these process inside
debuggers.

Previously, we would have the process utility module with knowledge about
an Avocado internal exception that flags that a test should exist as SKIP
in these circunstances. Now, to avoid these prior knowledge and layer
violation, let the user of the API register which exception should be
raised when such situations occur.

A more complex version of this change, deemed unnecessary for now,
would be to register callbacks which could, in turn, raise those exceptions.

Signed-off-by: Cleber Rosa <crosa@redhat.com>